+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `EdgeDB.ConfigMemory.new/1` to create a new instance of `t:EdgeDB.ConfigMemory.t/0`.
 - support for `Elixir v1.16`.
 - support for branches.
+- support for server name (SNI) passing to TLS connection.
 
 ### Changed